feat: Stage in step to use DAPA url to get the queries.  (#312)
* feat: adding support for dapa get url (tests are failing though)

* fix: update test dapa url
