+++ cumulus_lambda_functions/stage_in_out/search_granules_unity.py
@@ -2,6 +2,8 @@ import json
 import logging
 import os
 
+from cumulus_lambda_functions.uds_api.web_service_constants import WebServiceConstants
+
 from cumulus_lambda_functions.stage_in_out.dapa_client import DapaClient
 from cumulus_lambda_functions.stage_in_out.search_granules_abstract import SearchGranulesAbstract
 
@@ -24,6 +26,7 @@ class SearchGranulesUnity(SearchGranulesAbstract):
         self.__date_to = ''
         self.__limit = 1000
         self.__verify_ssl = True
+        self.__api_base_prefix = WebServiceConstants.API_PREFIX
 
     def __set_props_from_env(self):
         missing_keys = [k for k in [self.COLLECTION_ID_KEY] if k not in os.environ]
@@ -38,12 +41,13 @@ class SearchGranulesUnity(SearchGranulesAbstract):
 
         self.__date_from = os.environ.get(self.DATE_FROM_KEY, '')
         self.__date_to = os.environ.get(self.DATE_TO_KEY, '')
+        self.__api_base_prefix = os.environ.get('API_PREFIX', WebServiceConstants.API_PREFIX)
         self.__verify_ssl = os.environ.get(self.VERIFY_SSL_KEY, 'TRUE').strip().upper() == 'TRUE'
         return self
 
     def search(self, **kwargs) -> str:
         self.__set_props_from_env()
-        dapa_client = DapaClient().with_verify_ssl(self.__verify_ssl)
+        dapa_client = DapaClient(self.__api_base_prefix).with_verify_ssl(self.__verify_ssl)
         granules_result = dapa_client.get_all_granules(self.__collection_id, self.__limit, self.__date_from, self.__date_to)
         granules_result = granules_result
         return json.dumps(granules_result)
